pocs/qa-chatbot/requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25==0.2.2
pypdf>=4.0.0
python-docx>=1.0.0
rich>=13.0.0
numpy>=1.24.0
pytest==8.3.5
